ObjectenApiPlugin now supports DTOs
