+++ zgw/objecten-api/src/main/kotlin/com/ritense/objectenapi/client/ObjectRequest.kt
@@ -18,7 +18,7 @@ package com.ritense.objectenapi.client
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.databind.JsonNode
-import com.ritense.objectenapi.client.typed.TypedObjectRequest
+import com.ritense.objectenapi.client.dto.TypedObjectRequest
 import java.net.URI
 import java.util.UUID
